Finised recipe editing

+++ BuddhaBowls/ViewModels/NewRecipeVM.cs
@@ -15,8 +15,8 @@ namespace BuddhaBowls
 {
     public class NewRecipeVM : INotifyPropertyChanged
     {
-        private ModelContainer _models;
-        private List<IItem> _availableItems;
+        protected ModelContainer _models;
+        protected List<IItem> _availableItems;
         // INotifyPropertyChanged event and method
         public event PropertyChangedEventHandler PropertyChanged;
 
@@ -30,7 +30,6 @@ namespace BuddhaBowls
         #region Content Binders
 
         public string Name { get; set; }
-        public List<Recipe> RecipeItems { get; set; }
 
         private ObservableCollection<IItem> _ingredients;
         public ObservableCollection<IItem> Ingredients
@@ -93,13 +92,87 @@ namespace BuddhaBowls
             }
         }
 
+        public Visibility BatchVisibility { get; set; }
+        public Visibility MenuVisibility { get; set; }
+
+        private bool _isBatch;
+        public bool IsBatch
+        {
+            get
+            {
+                return _isBatch;
+            }
+            set
+            {
+                _isBatch = value;
+                BatchVisibility = _isBatch ? Visibility.Visible : Visibility.Hidden; 
+                MenuVisibility = _isBatch ? Visibility.Hidden : Visibility.Visible;
+                NotifyPropertyChanged("BatchVisibility");
+                NotifyPropertyChanged("MenuVisibility");
+            }
+        }
+
+        private string _header;
+        public string Header
+        {
+            get
+            {
+                return _header;
+            }
+            set
+            {
+                _header = value;
+                NotifyPropertyChanged("Header");
+            }
+        }
+
+        private string _errorMessage;
+        public string ErrorMessage
+        {
+            get
+            {
+                return _errorMessage;
+            }
+            set
+            {
+                _errorMessage = value;
+                NotifyPropertyChanged("ErrorMessage");
+            }
+        }
+
+        private ObservableCollection<FieldSetting> _fieldsCollection;
+        public ObservableCollection<FieldSetting> FieldsCollection
+        {
+            get
+            {
+                return _fieldsCollection;
+            }
+            set
+            {
+                _fieldsCollection = value;
+                NotifyPropertyChanged("FieldsCollection");
+            }
+        }
+
+        private string _modalTitle;
+        public string ModalTitle
+        {
+            get
+            {
+                return _modalTitle;
+            }
+            set
+            {
+                _modalTitle = value;
+                NotifyPropertyChanged("ModalTitle");
+            }
+        }
         #endregion
 
         #region ICommand Properties and Can Execute
 
         public ICommand AddItemCommand { get; set; }
         public ICommand RemoveItemCommand { get; set; }
-        public ICommand ResetCommand { get; set; }
         public ICommand SaveCommand { get; set; }
         public ICommand CancelCommand { get; set; }
         public ICommand ModalOkCommand { get; set; }
@@ -121,28 +194,41 @@ namespace BuddhaBowls
             }
         }
 
+        public bool SaveCanExecute
+        {
+            get
+            {
+                return string.IsNullOrEmpty(ErrorMessage);
+            }
+        }
+
         #endregion
 
-        public NewRecipeVM(ModelContainer models, MainViewModel mvm)
+        public NewRecipeVM(ModelContainer models, MainViewModel mvm, bool isBatch)
         {
             _models = models;
             ParentContext = mvm;
+            IsBatch = isBatch;
 
+            Header = "New " + (isBatch ? "Batch Recipe" : "Menu Item");
             _availableItems = new List<IItem>();
 
             AddItemCommand = new RelayCommand(AddItem);
             RemoveItemCommand = new RelayCommand(RemoveItem, x => RemoveCanExecute);
-            ResetCommand = new RelayCommand(ResetList);
             SaveCommand = new RelayCommand(Save);
             CancelCommand = new RelayCommand(Cancel);
             ModalOkCommand = new RelayCommand(ModalOk, x => ModalOkCanExecute);
             ModalCancelCommand = new RelayCommand(ModalCancel);
+
+            InitFieldsCollection();
+            Refresh();
         }
 
         #region ICommand Helpers
 
         private void AddItem(object obj)
         {
+            ModalTitle = "Add Ingredient";
             ParentContext.ModalContext = this;
             ModalVisibility = Visibility.Visible;
         }
@@ -155,8 +241,18 @@ namespace BuddhaBowls
 
         private void Save(object obj)
         {
-            // save recipe
-            ParentContext.DeleteTempTab();
+            Recipe newRecipe = new Recipe();
+            ErrorMessage = ParentContext.ObjectFromFields(ref newRecipe, FieldsCollection, true);
+            newRecipe.IsBatch = IsBatch;
+            newRecipe.ItemList = Ingredients.ToList();
+
+            if (string.IsNullOrEmpty(ErrorMessage))
+            {
+                newRecipe.Insert();
+                _models.Recipes.Add(newRecipe);
+                ParentContext.RecipeTab.RefreshList();
+                ParentContext.DeleteTempTab();
+            }
         }
 
         private void Cancel(object obj)
@@ -166,6 +262,7 @@ namespace BuddhaBowls
 
         private void ModalOk(object obj)
         {
+            ItemToAdd.Count = 0;
             _availableItems.Add(ItemToAdd);
             Refresh();
             ModalVisibility = Visibility.Hidden;
@@ -181,11 +278,38 @@ namespace BuddhaBowls
 
         #region Initializers
 
+        protected void InitFieldsCollection(IItem item = null)
+        {
+            if(item == null)
+            {
+                item = new Recipe();
+            }
+            if(IsBatch)
+            {
+                FieldsCollection = new ObservableCollection<FieldSetting>(new List<FieldSetting>()
+                {
+                    new FieldSetting("Name") { Value = item.Name },
+                    new FieldSetting("Category") { Value = item.Category },
+                    new FieldSetting("RecipeUnit") { Value = item.RecipeUnit },
+                    new FieldSetting("RecipeUnitConversion") { Value = item.RecipeUnitConversion.ToString() },
+                    new FieldSetting("Count") { Value = item.Count.ToString() },
+                });
+            }
+            else
+            {
+                FieldsCollection = new ObservableCollection<FieldSetting>(new List<FieldSetting>()
+                {
+                    new FieldSetting("Name") { Value = item.Name },
+                    new FieldSetting("Price") { Value = ((Recipe)item).Price.ToString() },
+                });
+            }
+        }
+
         #endregion
 
         #region Update UI
 
-        private void Refresh()
+        protected void Refresh()
         {
             if (_availableItems == null)
             {
@@ -199,6 +323,17 @@ namespace BuddhaBowls
                                                                                 .Where(x => !_availableItems.Select(y => y.Id).Contains(x.Id))
                                                                                 .OrderBy(x => x.Name));
             }
+        }
+
+        public void ClearErrors()
+        {
+            ErrorMessage = "";
+            foreach (FieldSetting field in FieldsCollection)
+            {
+                field.Error = 0;
+            }
+
+            NotifyPropertyChanged("FieldsCollection");
         }
 
         #endregion
